refactor video component
